server/course/api/serializers.py
from rest_framework import serializers
from course import models


class CourseSerializer(serializers.ModelSerializer):
    class Meta:
        model = models.Course
        fields = ('id', 'name', 'description')


